move TagsFrom tests to typegenTypes
